requirements-dev.txt
#Requirements for development.

flake8
ipython
mock
sphinx
coveralls
django-debug-toolbar
